+++ index.html
@@ -5,10 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <link rel="stylesheet" href="/css/style.css">
-    <script defer type="module" src="script/grid.js"></script>
-    <script defer type="module" src="script/rolling.js"></script>
-    <script defer type="module" src="script/topic.js"></script>
-    <script defer src="script/header.js"></script>
+    <script type="module" src="script/app.js"></script>
 </head>
 <body>
     <div class="news_stand_wrapper">
