+++ script/app.js
@@ -1,0 +1,11 @@
+import gridInit from './main/grid.js';
+import headerInit from './main/header.js';
+import rollingInit from './main/rolling.js';
+import getTopicInit from './main/topic.js';
+
+( () => {
+  gridInit();
+  headerInit();
+  rollingInit();
+  getTopicInit();
+})();
