+++ script/util/util.js
